make zip_same_timespan public
